move NotFound from pages
