+++ sample-apps/delivery-customizations-sample-app/web/frontend/components/NotFound.jsx
@@ -1,7 +1,8 @@
 import { Card, EmptyState, Page } from "@shopify/polaris";
 import { notFoundImage } from "../assets";
 
-export default function NotFound() {
+export function NotFound() {
+  console.log("Heyooo");
   return (
     <Page>
       <Card>
