+++ sample-apps/delivery-customizations-sample-app/web/frontend/components/index.js
@@ -1,4 +1,5 @@
-export { CustomizationForm } from './CustomizationForm'
-export {CustomizationPageLayout} from './CustomizationPageLayout'
-export {ErrorsBanner} from './ErrorsBanner'
+export { CustomizationForm } from "./CustomizationForm";
+export { CustomizationPageLayout } from "./CustomizationPageLayout";
+export { ErrorsBanner } from "./ErrorsBanner";
+export { NotFound } from "./NotFound";
 export * from "./providers";
